--- lib/model/test.txt
@@ -1,1 +1,0 @@
-titjgiorefnieoz
