--- lib/routes.dart
@@ -1,1 +1,0 @@
-enum Routes { Connection, Register, Navigation, StrollDetails, ForgotPassword }
