--- lib/screens/calendar.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Calendar extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("calendar"),
-    );
-  }
-}
