--- lib/screens/message.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-
-class Message extends StatefulWidget {
-  @override
-  _MessageState createState() => _MessageState();
-}
-
-class _MessageState extends State<Message> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Text("message"),
-    );
-  }
-}
