+++ pubspec.yaml
@@ -23,16 +23,9 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^0.7.0 
-  firebase_database: ^6.0.0
-  firebase_auth: ^0.20.1
-  cloud_firestore: ^0.16.0
-  email_validator: '^1.0.5'
-  intl: ^0.16.0
-  google_maps_flutter: ^2.0.1
-  map: ^0.3.0
-  cached_network_image: ^2.5.0
-  table_calendar: ^2.3.3
+  firebase_core: ^0.5.1 
+  firebase_database: ^4.2.0
+  firebase_auth: ^0.18.2
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -55,8 +48,10 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  assets:
-     - assets/
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
